Added Action and Gif Objects
Added Action: object: Given a name and an evemtListener event it will trigger activate on the event. Designed to be used with when: statement

Added gif: object: Given a gif directory will play all images in said directory in sequential order. Has the following properties:

x: the x position
y: the y position
src: the .gif directory
sticky: whether it interacts with scrolling
frame: the frame the gif will be initialized at
speed: the rate of which frames are shown

The gif format is designed as below:
Gifs are a directory with .pgi image files named 1, 2, 3, ... n for n frames. The gif will be displayed in numerical order.